refactor: Implemented relationship in URI
