+++ src/event/repositories/reservation.py
@@ -1,4 +1,5 @@
 from typing import Any
+from uuid import UUID
 from sqlalchemy import select
 from src.core.db import DbSession
 from src.event.models.reservation import Reservation
@@ -6,7 +7,10 @@ from src.event.models.reservation import Reservation
 
 class ReservationRepository:
     @staticmethod
-    def create(db: DbSession, serialized_data: dict[str, Any]) -> Reservation:
+    def create(
+        db: DbSession,
+        serialized_data: dict[str, Any]
+    ) -> Reservation:
         reservation = Reservation(**serialized_data)
 
         db.add(reservation)
@@ -17,9 +21,22 @@ class ReservationRepository:
         return reservation
 
     @staticmethod
-    def get_reservations(db: DbSession):
+    def get_reservations(
+        db: DbSession,
+        event_id: UUID,
+    ):
         result = db.scalars(
             select(Reservation)
+            .where(Reservation.event_id == event_id)
+        )
+
+        return result
+
+    @staticmethod
+    def get_reservations_by_user(db: DbSession, user_id: UUID):
+        result = db.scalars(
+            select(Reservation)
+            .where(Reservation.user_id == user_id)
         )
 
         return result
