+++ src/payment/enums.py
@@ -1,9 +1,9 @@
-from enum import Enum
+# from enum import Enum
 
 
-class TransactionPaymentStatuses(Enum):
-    COMPLETED = "transaction_payment_status_completed"
-    REFUNDED = "transaction_payment_status_refunded"
+# class TransactionPaymentStatuses(Enum):
+#     COMPLETED = "transaction_payment_status_completed"
+#     REFUNDED = "transaction_payment_status_refunded"
 
 
-TRANSACTION_PAYMENT_STATUS_TYPE = "transaction_payment_status_type"
+# TRANSACTION_PAYMENT_STATUS_TYPE = "transaction_payment_status_type"
